+++ apps/web/src/components/ui/buttons/FileUploadButton.tsx
@@ -32,13 +32,10 @@ export const FileUploadButton: React.FC<FileUploadButtonProps> = ({
   };
   
   const handleChange = React.useCallback((event: React.ChangeEvent<HTMLInputElement>) => {
-    console.log('FileUploadButton: handleChange called');
     
     // Process file upload
     const files = event.target.files;
-    console.log('FileUploadButton: Files selected', files);
     if (files && files.length > 0) {
-      console.log('FileUploadButton: Calling onChange with event');
       onChange(event);
     }
