+++ apps/web/src/hooks/execution/index.ts
@@ -1,16 +1,8 @@
 /**
- * Execution hooks - Modular execution management
+ * Execution hook exports
  * 
- * These hooks provide a clean separation of concerns for execution:
- * - useExecutionState: Pure state management
- * - useExecutionUI: UI concerns and formatting
- * - useExecutionV2: Combined hook for easy usage
+ * This module is kept for backward compatibility.
+ * New code should use useExecution from the parent hooks module.
  */
 
-export * from './useExecutionState';
-export * from './useExecutionUI';
-export * from './useExecutionV2';
-export * from './useDiagramRunner';
-
-// Re-export the main hook as default
-export { useExecutionV2 as default } from './useExecutionV2';
+// Empty - all execution functionality moved to useExecution hook
