+++ apps/web/src/types/domain/handle.ts
@@ -1,14 +1,9 @@
 import { HandleID, NodeID } from '../branded';
 import { DataType, HandlePosition } from '../primitives';
 
-/**
- * Handle direction - aligned terminology
- */
+
 export type HandleDirection = 'input' | 'output';
 
-/**
- * Pure domain handle - framework-agnostic port definition
- */
 export interface DomainHandle {
   id: HandleID;
   nodeId: NodeID;
@@ -21,26 +16,17 @@ export interface DomainHandle {
   maxConnections?: number;
 }
 
-/**
- * Input handle specification
- */
 export interface InputHandle extends DomainHandle {
   direction: 'input';
   required?: boolean;
   defaultValue?: unknown;
 }
 
-/**
- * Output handle specification
- */
 export interface OutputHandle extends DomainHandle {
   direction: 'output';
   dynamic?: boolean;
 }
 
-/**
- * Handle compatibility checker
- */
 export function areHandlesCompatible(
   source: DomainHandle,
   target: DomainHandle
@@ -60,16 +46,10 @@ export function areHandlesCompatible(
   return source.dataType === target.dataType;
 }
 
-/**
- * Generate handle ID from node ID and handle name
- */
 export function createHandleId(nodeId: NodeID, handleName: string): HandleID {
   return `${nodeId}:${handleName}` as HandleID;
 }
 
-/**
- * Parse handle ID into node ID and handle name
- */
 export function parseHandleId(handleId: HandleID): { nodeId: NodeID; handleName: string } {
   const [nodeId, ...handleNameParts] = handleId.split(':');
   return {
@@ -78,16 +58,10 @@ export function parseHandleId(handleId: HandleID): { nodeId: NodeID; handleName:
   };
 }
 
-/**
- * Validate handle ID format
- */
 export function isValidHandleIdFormat(value: string): boolean {
   return /^[a-zA-Z0-9_-]+:[a-zA-Z0-9_-]+$/.test(value);
 }
 
-/**
- * Validate node ID format
- */
 export function isValidNodeIdFormat(value: string): boolean {
   return /^[a-zA-Z0-9_-]+$/.test(value);
 }
