+++ apps/web/src/types/domain/node.ts
@@ -1,10 +1,6 @@
 import {NodeID, PersonID} from '../branded';
 import { Vec2, NodeKind, ConditionType, PersonForgettingStrategy, JobLanguage, DBOperation, DBSubType, NotionOperation, ProcessType } from '../primitives';
 
-/**
- * Pure domain node - framework-agnostic representation
- * No React Flow or UI framework dependencies
- */
 export interface DomainNode {
   id: NodeID;
   type: NodeKind;
@@ -12,18 +8,12 @@ export interface DomainNode {
   data: Record<string, unknown>;
 }
 
-/**
- * Node with typed data for specific node types
- */
 export interface TypedDomainNode<T extends NodeKind, D extends Record<string, unknown> = Record<string, unknown>>
   extends DomainNode {
   type: T;
   data: D;
 }
 
-/**
- * Start node domain model
- */
 export interface StartNodeData {
   label?: string;
   customData: Record<string, unknown>;
@@ -33,9 +23,6 @@ export interface StartNodeData {
 
 export type StartNode = TypedDomainNode<NodeKind, StartNodeData> & { type: 'start' };
 
-/**
- * Condition node domain model
- */
 export interface ConditionNodeData {
   label?: string;
   conditionType?: ConditionType;
@@ -47,9 +34,6 @@ export interface ConditionNodeData {
 
 export type ConditionNode = TypedDomainNode<NodeKind, ConditionNodeData> & { type: 'condition' };
 
-/**
- * Person job node domain model
- */
 export interface PersonJobNodeData {
   label?: string;
   person?: PersonID;
@@ -62,9 +46,6 @@ export interface PersonJobNodeData {
 
 export type PersonJobNode = TypedDomainNode<NodeKind, PersonJobNodeData> & { type: 'person_job' };
 
-/**
- * Endpoint node domain model
- */
 export interface EndpointNodeData {
   label?: string;
   saveToFile?: boolean;
@@ -74,9 +55,6 @@ export interface EndpointNodeData {
 
 export type EndpointNode = TypedDomainNode<NodeKind, EndpointNodeData> & { type: 'endpoint' };
 
-/**
- * DB node domain model
- */
 export interface DBNodeData {
   label?: string;
   file?: string;
@@ -90,9 +68,6 @@ export interface DBNodeData {
 
 export type DBNode = TypedDomainNode<NodeKind, DBNodeData> & { type: 'db' };
 
-/**
- * Job node domain model
- */
 export interface JobNodeData {
   label?: string;
   codeType: JobLanguage;
@@ -102,9 +77,6 @@ export interface JobNodeData {
 
 export type JobNode = TypedDomainNode<NodeKind, JobNodeData> & { type: 'job' };
 
-/**
- * User response node domain model
- */
 export interface UserResponseNodeData {
   label?: string;
   prompt: string;
@@ -114,9 +86,6 @@ export interface UserResponseNodeData {
 
 export type UserResponseNode = TypedDomainNode<NodeKind, UserResponseNodeData> & { type: 'user_response' };
 
-/**
- * Notion node domain model
- */
 export interface NotionNodeData {
   label?: string;
   operation: NotionOperation | 'create' | 'update';
@@ -130,9 +99,6 @@ export interface NotionNodeData {
 
 export type NotionNode = TypedDomainNode<NodeKind, NotionNodeData> & { type: 'notion' };
 
-/**
- * Person batch job node domain model
- */
 export interface PersonBatchJobNodeData {
   label?: string;
   person?: string;
@@ -146,9 +112,6 @@ export interface PersonBatchJobNodeData {
 
 export type PersonBatchJobNode = TypedDomainNode<NodeKind, PersonBatchJobNodeData> & { type: 'person_batch_job' };
 
-/**
- * Union of all domain node types
- */
 export type DiagramNode = 
   | StartNode
   | ConditionNode
@@ -160,9 +123,6 @@ export type DiagramNode =
   | NotionNode
   | PersonBatchJobNode;
 
-/**
- * Type guard for domain nodes
- */
 export function isDomainNode(obj: unknown): obj is DomainNode {
   return (
     obj !== null &&
@@ -174,9 +134,6 @@ export function isDomainNode(obj: unknown): obj is DomainNode {
   );
 }
 
-/**
- * Type guard for specific node types
- */
 export function isNodeOfKind<T extends NodeKind>(
   node: DomainNode,
   type: T
