+++ apps/web/src/types/domain/person.ts
@@ -1,18 +1,9 @@
 import { PersonID } from '../branded';
 
-/**
- * LLM service types
- */
 export type LLMService = 'openai' | 'claude' | 'gemini' | 'groq' | 'grok';
 
-/**
- * Forgetting mode for person memory management
- */
 export type ForgettingMode = 'no_forget' | 'on_every_turn' | 'upon_request';
 
-/**
- * Pure domain person - represents a configured LLM instance with memory
- */
 export interface DomainPerson {
   id: PersonID;
   name: string;
@@ -27,18 +18,12 @@ export interface DomainPerson {
   forgettingMode?: ForgettingMode;
 }
 
-/**
- * Person memory configuration
- */
 export interface PersonMemoryConfig {
   maxMessages?: number;
   ttlSeconds?: number;
   autoCleanup?: boolean;
 }
 
-/**
- * Person with extended configuration
- */
 export interface ExtendedPerson extends DomainPerson {
   memoryConfig?: PersonMemoryConfig;
   metadata?: {
@@ -48,9 +33,6 @@ export interface ExtendedPerson extends DomainPerson {
   };
 }
 
-/**
- * Type guard for domain person
- */
 export function isDomainPerson(obj: unknown): obj is DomainPerson {
   return (
     obj !== null &&
@@ -62,9 +44,6 @@ export function isDomainPerson(obj: unknown): obj is DomainPerson {
   );
 }
 
-/**
- * Default person configuration
- */
 export const DEFAULT_PERSON_CONFIG: Partial<DomainPerson> = {
   temperature: 0.7,
   maxTokens: 2000,
@@ -74,9 +53,6 @@ export const DEFAULT_PERSON_CONFIG: Partial<DomainPerson> = {
   forgettingMode: 'no_forget'
 };
 
-/**
- * Create a new person with defaults
- */
 export function createPerson(
   id: PersonID,
   name: string,
