--- apps/web/src/types/legacy/index.ts
@@ -1,11 +1,0 @@
-import type React from 'react';
-
-// Re-export ReactFlow types for backward compatibility
-export type { Connection, Edge, ReactFlowInstance } from '@xyflow/react';
-
-// ArrowData type for backward compatibility
-export type ArrowData = {
-  label?: string;
-  style?: React.CSSProperties;
-};
-
